fix: enums
- Ajout de l'enum menu dans le dossiers enums.
