+++ lib/enums/enums_export.dart
@@ -4,6 +4,9 @@ library;
 /// DOCUMENT FILTERED TYPE
 export 'document_filtered_type_enum.dart';
 
+/// MENU TEXT TYPE
+export 'menu_text_type_enum.dart';
+
 /// SCREEN MODE
 export 'screen_mode_enum.dart';
