+++ lib/server_nestjs/data/models/delete_params.dart
