+++ lib/server_nestjs/data/repositories/in_memory_api_nestjs_repository.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:flutter/cupertino.dart';
 // import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
-import 'package:observer_core/features/server_nestjs/feature_server_nestjs_export.dart';
+import 'package:observer_core/server_nestjs/feature_server_nestjs_export.dart';
 import 'package:retrofit/dio.dart';
 
 /// Class a instancier ou injecter dans un [GetIt] afin d'être utiliser <br>
