+++ lib/server_nestjs/data/sources/abstractions/abstract_server_nestjs_source.dart
@@ -1,4 +1,4 @@
-import 'package:observer_core/features/server_nestjs/feature_server_nestjs_export.dart';
+import 'package:observer_core/server_nestjs/feature_server_nestjs_export.dart';
 import 'package:retrofit/dio.dart';
 
 /// Détails des méthodes à implementer :
